backend/requirements.txt
Django==4.2.10
psycopg2>=2.8.6,<2.9
django-filter==23.5
djangorestframework==3.14.0
